# Conflicts:
#	src/main/java/fi/vaylavirasto/sillari/model/RouteModel.java
#	src/main/resources/graphql/schema.graphqls
#	ui/src/graphql/RouteQuery.ts
#	ui/src/interfaces/IRoute.ts

+++ src/main/java/fi/vaylavirasto/sillari/model/TransportModel.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class TransportModel {
     private long id;
     private long permitId;
-    private long routeId;
+    private Integer routeId;
     private String name;
     private Double height;
     private Double width;
